Safari-611.1.10/LayoutTests/platform/ipad/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-tokenization-innerheight-innerwidth-expected.txt


FAIL "innerwidth=401" should set width of opened window assert_equals: expected 401 but got 800
FAIL " innerwidth = 401" should set width of opened window assert_equals: expected 401 but got 800
FAIL "innerwidth==401" should set width of opened window assert_equals: expected 401 but got 800
FAIL "\ninnerwidth= 401" should set width of opened window assert_equals: expected 401 but got 800
FAIL ",innerwidth=401,," should set width of opened window assert_equals: expected 401 but got 800
FAIL "INNERWIDTH=401" should set width of opened window assert_equals: expected 401 but got 800
FAIL "innerWidth=401" should set width of opened window assert_equals: expected 401 but got 800
FAIL "innerheight=402" should set height of opened window assert_equals: expected 402 but got 600
FAIL " innerheight = 402" should set height of opened window assert_equals: expected 402 but got 600
FAIL "innerheight==402" should set height of opened window assert_equals: expected 402 but got 600
FAIL "\ninnerheight= 402" should set height of opened window assert_equals: expected 402 but got 600
FAIL ",innerheight=402,," should set height of opened window assert_equals: expected 402 but got 600
FAIL "INNERHEIGHT=402" should set height of opened window assert_equals: expected 402 but got 600
FAIL "innerHeight=402" should set height of opened window assert_equals: expected 402 but got 600
